notifications implemented, tests separated into different files, still more notification tests to write
