+++ server/controllers/index.js
@@ -1,9 +1,11 @@
 const reading = require('./reading');
 const node = require('./node');
 const user = require('./user');
+const notification = require('./notification');
 
 module.exports = {
   reading,
   node,
   user,
+  notification
 };
